refactor(spectral tools): trackviewtype -> trackviewtypeattachment
