+++ src/au3wrap/internal/trackviewtypeattachment.h
