+++ src/trackedit/internal/au3/au3tracksinteraction.cpp
@@ -23,7 +23,7 @@
 #include "au3wrap/internal/domconverter.h"
 #include "au3wrap/internal/trackcolor.h"
 #include "au3wrap/internal/trackrulertypeattachment.h"
-#include "au3wrap/internal/trackviewtype.h"
+#include "au3wrap/internal/trackviewtypeattachment.h"
 #include "au3wrap/internal/wxtypes_convert.h"
 #include "au3wrap/au3types.h"
